tests/legacy_tests/users/requirements.txt
# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# HTTP testing
requests>=2.31.0
httpx>=0.24.0

# Database testing
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Data validation and serialization
pydantic>=2.0.0
email-validator>=2.0.0

# Date/time handling
python-dateutil>=2.8.0

# Password hashing (for validation tests)
argon2-cffi>=23.0.0

# Test utilities
faker>=19.0.0
factory-boy>=3.3.0

# Performance testing
pytest-benchmark>=4.0.0

# Parallel test execution
pytest-xdist>=3.3.0

# Test reporting
pytest-html>=3.2.0
allure-pytest>=2.13.0

# Development utilities
pytest-watch>=4.2.0
pytest-sugar>=0.9.0
